Neopixel ambient light control via HTTP POST.
Send an HTTP POST request to http://hostname:8000/light with query parameters r=<0..255>, g=<0..255>, b=<0.255> and optionally time=<seconds> to set the light color.

Transitions nicely between different colors: When transitioning to a lighter color, the new color fades in from the center; when transitioning to a darker color, the new color fades in from the ends towards the center. Multiple concurrent transitions overlay in the order they were submitted.