Added simple web interface.
Extended the simple embedded web server to be able to deliver small files requested via HTTP GET, which must be stored in the 'web' subdirectory of the directory containing the ambientlight.py script file.

The web interface displays three buttons (white, warm, and off) that initiate transitions to the corresponding light color. It is set up to be added to the iPhone home screen as a standalone web application (with a pretty light bulb icon courtesy of https://icons8.com).

Also added support for 'GET /light', which returns the current target light color in JSON format. This isn't used yet, but the web interface may later use this interface to query the current light color in order to highlight the corresponding button (or something along those lines).