Added light color preview.
Makes the entire background of the app transition smoothly to represent the current light color. Queries the current light color per HTTP GET at startup and keeps it updated as the user changes it.

The display color is color-corrected prior to being rendered because the LED strip is very much tending towards blueish hues, which is already taken into consideration in the available color choices. Arguably, the color-correction should instead happen inside the ambient light controller (in both directions), not on client side.