Tweaked web app user interface.
The user interfaces fades in now and the buttons end with being slightly transparent so that the underlying light color preview shines through, which gives them a bit of the same hue as the light itself.

Changed JavaScript implementation to start modifying DOM not within the 'ready' event but intentionally delayed by a little. This reduces, but unfortunately doesn't completely avoid the screen background flashing white (for just a frame or so) between the all-black splash image and the all-black empty web app being displayed.

I have no idea how to completely suppress that white flash before the web app starts: It happens even when the HTML is reduced to just an empty <body style="background: black"> and even with an app cache manifest (which shows that it's not network fetch delays causing this).