Added LightManager to support high-level control.
While the low-level LightController transitions to a given light color, the LightManager has a concept of whether the light is 'on' or 'off' and supports cycling between a given set of light colors.

To support this, LightController now records a short history of the most recently set light colors, and implicitly ignores the attempt to set the current color again (even if it is set with a different transition time; a more comprehensive implementation might just overwrite the layer speed according to the new transition time).

Changed the HTTP interface of the "/light" URL to use PUT instead of POST for the old set-RGB-light-color interface and added a POST interface that takes a single parameter named "command" with supported values "on" (switch to most recently used 'on' color), "off" (switch to black) and "cycle" (switch to the next color in the configured cycle, or on if currently off).